Refactor verification logic and update Blockscout URLs
Moved contract verification processing logic from the cron route to a new processVerifications function in lib/solidity/verification-script.ts for better modularity and reuse. Updated the cron API route to use this new function and return its result. Commented out the Base Sepolia Blockscout URL in lib/blockscout.ts.
